Inno-SpaceProd-V2/requirements.txt
inno_utils @ git+https://dev.azure.com/SobeysInc/_git/Inno-Utils@master
databricks-connect==7.3.5
databricks_cli==0.11.0
cerberus==1.3
pytest==5.4.3
recommonmark==0.6.0
Sphinx==3.0.3
sphinx-autobuild==0.7.1
sphinx-rtd-theme==0.4.3
sphinx-markdown-tables==0.0.14
matplotlib==3.0.3
numpy==1.19.4
pandas==1.1.5
scikit-learn==0.24.1
scipy==1.4.1
pyarrow==3.0.0
python-box==3.4.1
pulp==2.6.0
pystan==2.19.1.1
python-dotenv==0.14.0
tensorflow==2.5.0
ortools==9.1.9490
six==1.15.0
dynamicTreeCut==0.1.1
mypy-extensions==0.4.3
statsmodels==0.13.2
openpyxl==3.0.9